remove outdated callers and replace with updated
