--- dev/vfm/call_vfm_sw.py
@@ -1,182 +1,0 @@
-from pathlib import Path
-
-import matplotlib.pyplot as plt
-import numpy as np
-
-from pyvale.vfm import (
-    BoundaryConditions,
-    ConstitutiveParameter,
-    Edge,
-    EdgeConditions,
-    EEdgeCondition,
-    ExperimentData,
-    HardeningLinear,
-    IdentificationConfig,
-    IdentificationPhase,
-    IsotropicVonMisesElastoplasticity,
-    run_identification,
-    SliceConfig,
-    SliceWiseForceReconstructionMetric,
-    SliceWiseIndependentLeastSquares,
-    SliceWiseSpatialParameterisation,
-    SpatialParameterisationKnown,
-    SpecimenGeometry,
-    SupportSlice,
-    VectorWeightedObjective,
-    VfmRegionOfInterest,
-)
-from pyvale.vfm.spatialparamslicewise import build_slice_partition
-
-
-INPUTS_PATH = Path(__file__).resolve().parent / "rob-data" / "wdbn4-vfm-input-data-260629-1530"
-SLICE_AXIS = "y"
-NUM_SLICES = 20
-PLOT_SLICE_PARTITION = True
-PLOT_SLICE_INDEX = 0
-
-
-def _print_coverage_diagnostic(slice_partition) -> None:
-    mean_coverage = float(np.mean(slice_partition.coverage_fractions))
-    min_slice_index = int(np.argmin(slice_partition.coverage_fractions))
-    min_slice_coverage = float(slice_partition.coverage_fractions[min_slice_index])
-    overall_coverage = float(np.sum(slice_partition.areas) / np.sum(slice_partition.geometric_areas))
-
-    print("Slice coverage diagnostic:")
-    print(f"  overall area coverage ratio: {overall_coverage:.6f}")
-    print(f"  mean slice coverage ratio:   {mean_coverage:.6f}")
-    print(f"  minimum slice coverage:      {min_slice_coverage:.6f} (slice {min_slice_index})")
-    print("  per-slice coverage ratios:")
-    for slice_index, coverage in enumerate(slice_partition.coverage_fractions):
-        print(f"    slice {slice_index:>2d}: {float(coverage):.6f}")
-
-
-def main() -> None:
-
-    # Prepare experiment data
-    specimen_geometry = SpecimenGeometry(
-        x=np.load(INPUTS_PATH / "x.npy"),
-        y=np.load(INPUTS_PATH / "y.npy"),
-        region_of_interest=VfmRegionOfInterest.from_yaml(INPUTS_PATH / "region_of_interest.yaml"),
-        thickness=0.8,
-        pixel_area=np.load(INPUTS_PATH / "pixel_area.npy"),
-    )
-
-    boundary_conditions = BoundaryConditions(
-        EdgeConditions(
-            min_x_edge=Edge(
-                EEdgeCondition.Free,
-                EEdgeCondition.Free,
-            ),
-            max_x_edge=Edge(
-                EEdgeCondition.Free,
-                EEdgeCondition.Free,
-            ),
-            min_y_edge=Edge(
-                EEdgeCondition.Fixed,
-                EEdgeCondition.Fixed,
-            ),
-            max_y_edge=Edge(
-                EEdgeCondition.Free,
-                EEdgeCondition.Traction,
-            ),
-        ),
-        np.load(INPUTS_PATH / "force.npy"),
-    )
-
-    experiment_data = ExperimentData(
-        np.load(INPUTS_PATH / "strain.npy"),
-        specimen_geometry,
-        boundary_conditions,
-        np.load(INPUTS_PATH / "time.npy"),
-    )
-
-    # Define constitutive parameters associated with constitutive law
-    constitutive_law = IsotropicVonMisesElastoplasticity(HardeningLinear())
-    parameter_map_size = np.array(specimen_geometry.x.shape)
-    parameters = {
-        "elastic_modulus": ConstitutiveParameter(
-            210_000, 150_000, 250_000, parameter_map_size
-        ),
-        "poissons_ratio": ConstitutiveParameter(
-            0.3, 0.2, 0.4, parameter_map_size
-        ),
-        "yield_strength": ConstitutiveParameter(
-            200, 100, 2000, parameter_map_size
-        ),
-        "hardening_modulus": ConstitutiveParameter(
-            3000, 1000, 15_000, parameter_map_size
-        ),
-    }
-
-    # Define shared spatial parameterisation supports
-    # Note for supports: shared object means shared evolution, separate objects mean independent ownership.
-    shared_slice_support = SupportSlice(
-        slice_config=SliceConfig(
-            axis=SLICE_AXIS,
-            num_slices=NUM_SLICES,
-        )
-    )
-
-    # slice_partition = build_slice_partition(
-    #         specimen_geometry,
-    #         slice_config=slice_config,
-    #         plot_diagnostic=True,
-    #         diagnostic_slice_index=PLOT_SLICE_INDEX,
-    #     )
-    # _print_coverage_diagnostic(slice_partition)
-
-    # Define identification phases
-    # Each phase has its own set of spatial parameterisations, metrics, objective function and optimiser. 
-    phases = [
-        # Phase 0
-        IdentificationPhase(
-            spatial_parameterisations={
-                "elastic_modulus": [SpatialParameterisationKnown()],
-                "poissons_ratio": [SpatialParameterisationKnown()],
-                "yield_strength": [
-                    SliceWiseSpatialParameterisation(
-                        support=shared_slice_support
-                    )
-                ],
-                "hardening_modulus": [
-                    SliceWiseSpatialParameterisation(
-                        support=shared_slice_support
-                    )
-                ],
-            },
-            metrics=[
-                SliceWiseForceReconstructionMetric(
-                    support=shared_slice_support
-                )
-            ],
-            objective_function=VectorWeightedObjective(),
-            optimiser=SliceWiseIndependentLeastSquares(),
-        )
-    ]
-
-    # Assemble the identification configuration
-    identification = IdentificationConfig(
-        constitutive_law=constitutive_law,
-        parameters=parameters,
-        phases=phases,
-    )
-
-    # Run the identification
-    vfm_result = run_identification(experiment_data, identification)
-    print(vfm_result)
-
-    fig, axes = plt.subplots(1, 2, figsize=(10, 4), constrained_layout=True)
-    for ax, param_name, title in zip(
-        axes,
-        ("yield_strength", "hardening_modulus"),
-        ("Yield Strength", "Hardening Modulus"),
-        strict=True,
-    ):
-        image = ax.imshow(vfm_result.parameter_maps[param_name], origin="lower", cmap="viridis")
-        ax.set_title(title)
-        fig.colorbar(image, ax=ax)
-    plt.show()
-
-
-if __name__ == "__main__":
-    main()
